Add functionality to install debian 9
This update changes the linux() function to allow switching based on linux distributions.
To simplify the execution it follows Gregor's advice to use the 'lsb_release' os command. 
This can further be extended to add functionality for RHEL systems by simply adding a case in the linux function. 
Debian 10 was not added since the Mongo site does not have a guide. This may be because debian 10 was released this month.  
If it is necessary to have Debian 10 it may work with a simply switch libcurl3 with libcurl4.