Migrate config ssh commands to key (gen or verify)
The creation and verification of keys should be housed within the key command.
Though the config does in fact use keys, the responsibility to handle them
should be controlled by commands related to keys.

The config ssh commands also used the EncryptFile class which lacked many
security and integrity features. The new key commands utilize the KeyHandler
class for checks and generation.